# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciamento Clinico",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "description": "Retorna uma lista de clientes",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Criar cliente",
                "description": "Cria um novo cliente",
                "operationId": "f1e86e5b79fedfa6593067977292b13b",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreClientRequest"
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exibir cliente",
                "description": "Retorna os dados de um cliente específico",
                "operationId": "4c3c7071ea1133f04bacdbc70f8aa230",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar cliente",
                "description": "Atualiza os dados de um cliente específico",
                "operationId": "7d0a5ae9e306858d30e04fc50265d80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateClientRequest"
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir cliente",
                "description": "Exclui um cliente específico",
                "operationId": "c929f14f421a80a896c6fe116ffdc317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente excluído"
                    }
                }
            }
        },
        "/api/exams": {
            "get": {
                "tags": [
                    "Exames"
                ],
                "summary": "Listar exames",
                "description": "Retorna uma lista de exames",
                "operationId": "a54e05ebd2b23c9f2329a2a483272bc3",
                "responses": {
                    "200": {
                        "description": "Lista de exames",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exam"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exames"
                ],
                "summary": "Criar exame",
                "description": "Cria um novo exame",
                "operationId": "fa2de658c3313746441306155597563e",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreExamRequest"
                },
                "responses": {
                    "201": {
                        "description": "Exame criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exams/{id}": {
            "get": {
                "tags": [
                    "Exames"
                ],
                "summary": "Exibir exame",
                "description": "Retorna os dados de um exame específico",
                "operationId": "5762ee3221f12d62693826bab5e66e2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do exame",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Exames"
                ],
                "summary": "Atualizar exame",
                "description": "Atualiza os dados de um exame específico",
                "operationId": "d8b5e491c38a96d9bd5f9787cff42ab1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateExamRequest"
                },
                "responses": {
                    "200": {
                        "description": "Exame atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exam"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exames"
                ],
                "summary": "Excluir exame",
                "description": "Exclui um exame específico",
                "operationId": "61fff9ddd42647f37ba3b48f3c0ac233",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do exame",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Exame excluído"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Cliente",
                "description": "Modelo de Cliente",
                "required": [
                    "name",
                    "cpf",
                    "birth_date",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "ID do cliente",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do cliente",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do cliente",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Data de nascimento do cliente",
                        "type": "string",
                        "format": "date"
                    },
                    "email": {
                        "description": "Email do cliente",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Telefone do cliente",
                        "type": "string"
                    },
                    "address": {
                        "description": "Endereço do cliente",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Exam": {
                "title": "Exame",
                "description": "Modelo de Exame",
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "description": "ID do exame",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do exame",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do exame",
                        "type": "string"
                    },
                    "price": {
                        "description": "Preço do exame",
                        "type": "number",
                        "format": "float"
                    },
                    "duration": {
                        "description": "Duração do exame",
                        "type": "string",
                        "format": "time"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
            "StoreClientRequest": {
                "description": "Dados para criar um cliente",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do cliente",
                                    "type": "string"
                                },
                                "cpf": {
                                    "description": "CPF do cliente",
                                    "type": "string"
                                },
                                "birth_date": {
                                    "description": "Data de nascimento do cliente",
                                    "type": "string",
                                    "format": "date"
                                },
                                "email": {
                                    "description": "Email do cliente",
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone": {
                                    "description": "Telefone do cliente",
                                    "type": "string"
                                },
                                "address": {
                                    "description": "Endereço do cliente",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreExamRequest": {
                "description": "Dados para criar um exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do exame",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Descrição do exame",
                                    "type": "string"
                                },
                                "price": {
                                    "description": "Preço do exame",
                                    "type": "number",
                                    "format": "float"
                                },
                                "duration": {
                                    "description": "Duração do exame",
                                    "type": "string",
                                    "format": "time"
                                },
                                "icon": {
                                    "description": "Ícone do exame",
                                    "type": "string"
                                },
                                "color": {
                                    "description": "Cor do exame",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateClientRequest": {
                "description": "Dados para atualizar um cliente",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do cliente",
                                    "type": "string"
                                },
                                "cpf": {
                                    "description": "CPF do cliente",
                                    "type": "string"
                                },
                                "birth_date": {
                                    "description": "Data de nascimento do cliente",
                                    "type": "string",
                                    "format": "date"
                                },
                                "email": {
                                    "description": "Email do cliente",
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone": {
                                    "description": "Telefone do cliente",
                                    "type": "string"
                                },
                                "address": {
                                    "description": "Endereço do cliente",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateExamRequest": {
                "description": "Dados para atualizar um exame",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do exame",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Descrição do exame",
                                    "type": "string"
                                },
                                "price": {
                                    "description": "Preço do exame",
                                    "type": "number",
                                    "format": "float"
                                },
                                "duration": {
                                    "description": "Duração do exame",
                                    "type": "string",
                                    "format": "time"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}